LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-model/grid-button-001-expected.txt


item 1
item 2

FAIL [data-expected-width] 1 assert_equals:
<span class="item1" data-expected-width="100">item 1</span>
width expected 100 but got 35
FAIL [data-expected-width] 2 assert_equals:
<span class="item2" data-expected-width="200">item 2</span>
width expected 200 but got 35
